backend/pyproject.toml
[project]
name = "onramp-backend"
version = "0.1.0"
description = "OnRamp - Azure Landing Zone Architect & Deployer API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aioodbc>=0.5.0",
    "azure-identity>=1.17.0",
    "azure-ai-projects>=1.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-subscription>=3.0.0",
    "msal>=1.28.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
